Agregando user service

+++ backend/src/main/java/com/project/c17567Java/Service/IUserService.java
@@ -1,13 +1,13 @@
 package com.project.c17567Java.Service;
 
-import com.project.c17567Java.Auth.Entity.User;
+import com.project.c17567Java.Auth.User.User;
 
 import java.util.List;
 
 public interface IUserService {
     public void saveUsuario(User user);
-    public void deleteUsuario(Long id);
-    public User findUsuario(Long id);
+    public void deleteUsuario(Integer id);
+    public User findUsuario(Integer id);
     public void editUsuario(User user);
     public List<User> getUsuarios();
 }
